Move package deps to internal add makefile
Moved all of the internal dependences to the standard internal directory
to clear out the way for ci configs and local build configs to live in
the build directory (which was already in use).  Replaced the compile
script with a makefile for more control of the build.  I'm pretty sure
that I broke docker compose but I will test later.  Also updated the
import paths.