Add dockerfiles and set up build and gcr push
Added all of the dockerfiles that we need for building the service
containers, clean makefiles, and update the build config create the
binaries and push them to gcr.  Moved dockerfiles into their own
directories during a failed attempt to use the grc orb but ended up
liking them that way so kept them there.