Update to reflect new app structure and add tags
Removed the server types in favor of a protocol configurable chat
server.  Added tagging support so we are using the tag environment
variables to set the new versions and just not overwriting latest.  K8s
was actually not picking up the overwrite of the tag for deployment.
Bake the message schema into the image and get rid of the run target
since it was previously updated to composer-up/composer-down.