Embed the healthcheck.
GCE ingress for GKE does not have the flexibility to 1) easily add
healthchecks that run on different ports and 2) healthchecks on the load
balancers cannot upgrade websocket connections which is what the
original route set for the root path.  This change swaps the root
path with a standard http response that the web check can use and puts
the websocket handler under /chat.